Renamed file, called the service handel properly
